summary-generation-handler/requirements.txt
redis==5.0.1
google-genai==1.55.0
